+++ database/migrations/2025_03_18_162748_create_product_variants_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class () extends Migration {
+return new class() extends Migration {
     /**
      * Run the migrations.
      */
@@ -16,8 +16,8 @@ return new class () extends Migration {
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->string('name');
             $table->string('sku')->nullable(); // SKU (Stock Keeping Unit)
-            $table->float('price');
-            $table->integer('compare_at_price')->nullable();
+            $table->decimal('price', 10, 2);
+            $table->decimal('compare_at_price', 10, 2)->nullable();
             $table->string('stock_status')->nullable(); // e.g., "inStock"
             $table->integer('stock_count')->default(0);
             $table->integer('weight')->nullable();
